# api.yaml
openapi: 3.0.3
info:
  title: Room Booking Service (Test Task)
  description: Сервис бронирования переговорок
  version: "1.0.0"
servers:
  - url: http://localhost:8080
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    BookingIdPath:
      name: bookingId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Идентификатор брони
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum:
                - INVALID_REQUEST
                - UNAUTHORIZED
                - NOT_FOUND
                - ROOM_NOT_FOUND
                - SLOT_NOT_FOUND
                - SLOT_ALREADY_BOOKED
                - BOOKING_NOT_FOUND
                - FORBIDDEN
                - SCHEDULE_EXISTS
                - INTERNAL_ERROR
            message:
              type: string
      example:
        error:
          code: INVALID_REQUEST
          message: invalid request
    InternalErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: INTERNAL_ERROR
            message:
              type: string
      example:
        error:
          code: INTERNAL_ERROR
          message: internal server error
...
    Room:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          nullable: true
        capacity:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
          nullable: true
...
    Booking:
      type: object
      required: [id, slotId, userId, status]
      properties:
        id:
          type: string
          format: uuid
        slotId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        status:
          type: string
          enum: [active, cancelled]
        conferenceLink:
          type: string
          format: uri
          nullable: true
        createdAt:
          type: string
          format: date-time
          nullable: true
    Pagination:
      type: object
      required: [page, pageSize, total]
      properties:
        page:
          type: integer
          description: Текущая страница (начиная с 1)
        pageSize:
          type: integer
          description: Размер страницы
        total:
          type: integer
          description: Общее количество записей
paths:
...
  /rooms/list:
    get:
      tags: [Rooms]
      summary: Список переговорок (admin и user)
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          description: Список переговорок
          content:
            application/json:
              schema:
                type: object
                properties:
                  rooms:
                    type: array
                    items:
                      $ref: '#/components/schemas/Room'
        '401':
          description: Не авторизован
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InternalErrorResponse' }
  /rooms/create:
    post:
      tags: [Rooms]
      summary: Создать переговорку (только admin)
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                description:
                  type: string
                  nullable: true
                capacity:
                  type: integer
                  nullable: true
      responses:
        '201':
          description: Переговорка создана
          content:
            application/json:
              schema:
                type: object
                properties:
                  room:
                    $ref: '#/components/schemas/Room'
        '400':
          description: Неверный запрос
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401':
          description: Не авторизован
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '403':
          description: Доступ запрещён (требуется роль admin)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InternalErrorResponse' }

  # ──────────────────────────────────────────────────────────────
  # Schedules
  # ──────────────────────────────────────────────────────────────
...
  /bookings/create:
    post:
      tags: [Bookings]
      summary: Создать бронь на слот (только user). Опционально — запросить ссылку на конференцию.
      description: >
        Доступно только роли user. Администратор не может создавать брони (403).
        `userId` берётся из JWT-токена, а не из тела запроса.
        Если временной слот находится в прошлом (start < now), возвращается 400 (INVALID_REQUEST).
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [slotId]
              properties:
                slotId:
                  type: string
                  format: uuid
                createConferenceLink:
                  type: boolean
                  default: false
      responses:
        '201':
          description: Бронь создана
          content:
            application/json:
              schema:
                type: object
                properties:
                  booking:
                    $ref: '#/components/schemas/Booking'
        '400':
          description: Неверный запрос
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401':
          description: Не авторизован
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '403':
          description: Доступ запрещён (бронирование доступно только роли user)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404':
          description: Слот не найден
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '409':
          description: Слот уже занят
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
              example:
                error:
                  code: SLOT_ALREADY_BOOKED
                  message: slot is already booked
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InternalErrorResponse' }
  /bookings/list:
    get:
      tags: [Bookings]
      summary: Список всех броней с пагинацией (только admin)
      description: >
        Доступно только роли admin.
        Поддерживает пагинацию через параметры `page` и `pageSize`.
        Оба параметра опциональны; значения по умолчанию: `page=1`, `pageSize=20`.
        Максимальное значение `pageSize` — 100.
      security: [{ BearerAuth: [] }]
      parameters:
        - name: page
          in: query
          description: Номер страницы (начиная с 1). По умолчанию 1.
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          description: Количество записей на странице. По умолчанию 20, максимум 100.
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Список всех броней
          content:
            application/json:
              schema:
                type: object
                properties:
                  bookings:
                    type: array
                    items:
                      $ref: '#/components/schemas/Booking'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '400':
          description: Неверный запрос (некорректные параметры пагинации)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401':
          description: Не авторизован
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '403':
          description: Доступ запрещён (только admin)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InternalErrorResponse' }
  /bookings/my:
    get:
      tags: [Bookings]
      summary: Список броней текущего пользователя (только user)
      description: >
        Доступно только роли user.
        Возвращает брони пользователя, чей `user_id` содержится в JWT-токене.
        Возвращаются только брони на будущие слоты (start >= now); брони на уже прошедшие слоты в ответ не включаются.
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          description: Список броней текущего пользователя
          content:
            application/json:
              schema:
                type: object
                properties:
                  bookings:
                    type: array
                    items:
                      $ref: '#/components/schemas/Booking'
        '401':
          description: Не авторизован
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '403':
          description: Доступ запрещён (только user)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InternalErrorResponse' }
  /bookings/{bookingId}/cancel:
    post:
      tags: [Bookings]
      summary: Отменить бронь (только своя бронь, только user)
      description: >
        Доступно только роли user. Пользователь может отменить только свою бронь.
        Операция **идемпотентна**: повторный вызов на уже отменённой брони не является ошибкой
        и возвращает 200 с актуальным состоянием брони (status: cancelled).
        При попытке отменить чужую бронь — 403.
        При отсутствии брони с указанным ID — 404.
      security: [{ BearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/BookingIdPath'
      responses:
        '200':
          description: Бронь отменена (или уже была отменена ранее)
          content:
            application/json:
              schema:
                type: object
                properties:
                  booking:
                    $ref: '#/components/schemas/Booking'
        '401':
          description: Не авторизован
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '403':
          description: Не своя бронь или роль не user
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
              example:
                error:
                  code: FORBIDDEN
                  message: cannot cancel another user's booking
        '404':
          description: Бронь не найдена
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InternalErrorResponse' }